Add build features in code

+++ video_maker/bots/cleaner.js
@@ -1,16 +1,43 @@
 const printer = require('./print');
+const {buildPath} = require('./path');
 const fs = require('fs');
 
 function main() {
     printer.print('>> Initializing cleaner bot');
     printer.sendInfo('bot-start::cleaner');
+    const path = buildPath('video_maker/temp');
+    printer.print('>> Verifying temp folders');
+    printer.sendInfo('bot-process::cleaner::verifying-folders');
+    createDataDirectories();
+    createTempDirectories(path);
     printer.sendInfo('bot-process::cleaner::cleaning-temp');
-    const tmp = fs.readdirSync('video_maker/temp');
+    const tmp = fs.readdirSync(path);
     tmp.forEach(e => {
-        clearDirectory(`video_maker/temp/${e}`)
+        clearDirectory(`${path}/${e}`)
     });
     printer.print('>>> Temp files cleaned');
     printer.sendInfo('bot-end::cleaner');
+}
+
+function createTempDirectories(path) {
+    if(!fs.existsSync(path)) fs.mkdirSync(path);
+    const dir = [
+        'background',
+        'foreground',
+        'music',
+        'slide'
+    ];
+    dir.forEach(e => {
+        if(!fs.existsSync(`${path}/${e}`)) {
+            fs.mkdirSync(`${path}/${e}`);
+        }
+    });
+}
+
+function createDataDirectories() {
+    const path = buildPath('video_maker/data');
+    if(!fs.existsSync(path)) fs.mkdirSync(path);
+    if(!fs.existsSync(`${path}/images`)) fs.mkdirSync(`${path}/images`);
 }
 
 function clearDirectory(path) {
@@ -19,7 +46,7 @@ function clearDirectory(path) {
         sub.forEach(e => {
             clearDirectory(`${path}/${e}`)
         });
-    } else {
+    } else if(!path.includes('.lock')) {
         fs.unlinkSync(path);
     }
 }
